Go ahead and submit the form and watch the router navigate to the new invoice.
Oh, one more thing. Route Components get a navigate prop. This version of the function is aware of its position in the hierarchy. This means you can navigate to relative paths the same way you can link to them.

Go ahead and remove the spot where we imported navigate and let’s use the prop instead: